created the alpha subdomain and edited the .htaccess
the subdomain alpha/ is now up and running. I transferred all the files from www/ to alpha/ and edited the htaccess to redirect to alpha/ when visiting harrywishyouwell.com.
I also added some bad bots blocking stuff and some protection to the htaccess.
edited the readme.md to take account of the changes

Signed-off-by: Ladislas de Toldi <ladislas@leka-smarttoys.com>